+++ rds-20140815/alibabacloud_rds20140815/models/__init__.py
@@ -152,9 +152,6 @@ from ._create_rcdeployment_set_response import CreateRCDeploymentSetResponse
 from ._create_rcdisk_request import CreateRCDiskRequest
 from ._create_rcdisk_response_body import CreateRCDiskResponseBody
 from ._create_rcdisk_response import CreateRCDiskResponse
-from ._create_rcimage_request import CreateRCImageRequest
-from ._create_rcimage_response_body import CreateRCImageResponseBody
-from ._create_rcimage_response import CreateRCImageResponse
 from ._create_rcnode_pool_request import CreateRCNodePoolRequest
 from ._create_rcnode_pool_shrink_request import CreateRCNodePoolShrinkRequest
 from ._create_rcnode_pool_response_body import CreateRCNodePoolResponseBody
@@ -885,9 +882,6 @@ from ._modify_dbinstance_spec_response import ModifyDBInstanceSpecResponse
 from ._modify_dbinstance_tderequest import ModifyDBInstanceTDERequest
 from ._modify_dbinstance_tderesponse_body import ModifyDBInstanceTDEResponseBody
 from ._modify_dbinstance_tderesponse import ModifyDBInstanceTDEResponse
-from ._modify_dbinstance_vector_support_status_request import ModifyDBInstanceVectorSupportStatusRequest
-from ._modify_dbinstance_vector_support_status_response_body import ModifyDBInstanceVectorSupportStatusResponseBody
-from ._modify_dbinstance_vector_support_status_response import ModifyDBInstanceVectorSupportStatusResponse
 from ._modify_dbnode_request import ModifyDBNodeRequest
 from ._modify_dbnode_shrink_request import ModifyDBNodeShrinkRequest
 from ._modify_dbnode_response_body import ModifyDBNodeResponseBody
@@ -946,9 +940,6 @@ from ._modify_parameter_response import ModifyParameterResponse
 from ._modify_parameter_group_request import ModifyParameterGroupRequest
 from ._modify_parameter_group_response_body import ModifyParameterGroupResponseBody
 from ._modify_parameter_group_response import ModifyParameterGroupResponse
-from ._modify_rcdisk_attribute_request import ModifyRCDiskAttributeRequest
-from ._modify_rcdisk_attribute_response_body import ModifyRCDiskAttributeResponseBody
-from ._modify_rcdisk_attribute_response import ModifyRCDiskAttributeResponse
 from ._modify_rcdisk_charge_type_request import ModifyRCDiskChargeTypeRequest
 from ._modify_rcdisk_charge_type_response_body import ModifyRCDiskChargeTypeResponseBody
 from ._modify_rcdisk_charge_type_response import ModifyRCDiskChargeTypeResponse
@@ -1744,9 +1735,6 @@ __all__ = [
     CreateRCDiskRequest,
     CreateRCDiskResponseBody,
     CreateRCDiskResponse,
-    CreateRCImageRequest,
-    CreateRCImageResponseBody,
-    CreateRCImageResponse,
     CreateRCNodePoolRequest,
     CreateRCNodePoolShrinkRequest,
     CreateRCNodePoolResponseBody,
@@ -2477,9 +2465,6 @@ __all__ = [
     ModifyDBInstanceTDERequest,
     ModifyDBInstanceTDEResponseBody,
     ModifyDBInstanceTDEResponse,
-    ModifyDBInstanceVectorSupportStatusRequest,
-    ModifyDBInstanceVectorSupportStatusResponseBody,
-    ModifyDBInstanceVectorSupportStatusResponse,
     ModifyDBNodeRequest,
     ModifyDBNodeShrinkRequest,
     ModifyDBNodeResponseBody,
@@ -2538,9 +2523,6 @@ __all__ = [
     ModifyParameterGroupRequest,
     ModifyParameterGroupResponseBody,
     ModifyParameterGroupResponse,
-    ModifyRCDiskAttributeRequest,
-    ModifyRCDiskAttributeResponseBody,
-    ModifyRCDiskAttributeResponse,
     ModifyRCDiskChargeTypeRequest,
     ModifyRCDiskChargeTypeResponseBody,
     ModifyRCDiskChargeTypeResponse,