+++ rds-20140815/alibabacloud_rds20140815/models/_delete_rcdisk_request.py
@@ -15,6 +15,8 @@ class DeleteRCDiskRequest(DaraModel):
         # This parameter is required.
         self.disk_id = disk_id
         # The region ID of the instance.
+        # 
+        # This parameter is required.
         self.region_id = region_id
 
     def validate(self):