+++ rds-20140815/alibabacloud_rds20140815/models/_stop_rcinstance_request.py
@@ -10,7 +10,6 @@ class StopRCInstanceRequest(DaraModel):
         force_stop: bool = None,
         instance_id: str = None,
         region_id: str = None,
-        stopped_mode: str = None,
     ):
         # Specifies whether to forcefully stop the instance. Valid values:
         # 
@@ -23,7 +22,6 @@ class StopRCInstanceRequest(DaraModel):
         self.instance_id = instance_id
         # The region ID.
         self.region_id = region_id
-        self.stopped_mode = stopped_mode
 
     def validate(self):
         pass
@@ -42,9 +40,6 @@ class StopRCInstanceRequest(DaraModel):
         if self.region_id is not None:
             result['RegionId'] = self.region_id
 
-        if self.stopped_mode is not None:
-            result['StoppedMode'] = self.stopped_mode
-
         return result
 
     def from_map(self, m: dict = None):
@@ -57,9 +52,6 @@ class StopRCInstanceRequest(DaraModel):
 
         if m.get('RegionId') is not None:
             self.region_id = m.get('RegionId')
-
-        if m.get('StoppedMode') is not None:
-            self.stopped_mode = m.get('StoppedMode')
 
         return self